Update return type of local qHash overload
The Qt 6 prototype returns size_t.

Pick-to: 6.1

+++ src/quicktemplates2/qquickdeferredexecute.cpp
@@ -50,7 +50,7 @@ namespace QtQuickPrivate {
 
 typedef QHash<uint, QQmlComponentPrivate::DeferredState *> DeferredStates;
 
-static inline uint qHash(QObject *object, const QString &propertyName)
+static inline size_t qHash(QObject *object, const QString &propertyName)
 {
     return ::qHash(object) + ::qHash(propertyName);
 }
